+++ turf-form.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Commercial Turf Rebate & AB 1572 Compliance</title>
+    <link rel="stylesheet" href="style.css" />
+    <meta name="api-base" content="https://nftapi.cyberwiz.io" />
+  </head>
+  <body>
+    <main class="container">
+      <h1>Commercial Turf Rebate & AB 1572 Compliance</h1>
+      <pre id="databaseDump" aria-live="polite"></pre>
+    </main>
+    <script type="module" src="dist/turf-form.js"></script>
+  </body>
+</html>
