+++ vite.config.js
@@ -4,9 +4,12 @@ export default defineConfig({
   envPrefix: ["VITE_", "MAPS_"],
   build: {
     rollupOptions: {
-      input: "src/main.js",
+      input: {
+        main: "src/main.js",
+        "turf-form": "src/turf-form.js",
+      },
       output: {
-        entryFileNames: "main.js",
+        entryFileNames: "[name].js",
         chunkFileNames: "[name].js",
         manualChunks: {
           maps: ["src/maps.js"],
